requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-huggingface>=0.1.0
langchain-text-splitters>=0.3.0
langgraph>=0.2.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
pypdf>=5.0.0
python-dotenv>=1.0.0
reportlab>=4.0.0
tavily-python>=0.5.0
